Base Account functionality completed

+++ src/main/java/org/example/simplemoneytransfer/restcontroller/BankController.java
@@ -6,9 +6,12 @@ import java.util.stream.Collectors;
 import org.example.simplemoneytransfer.dto.BankDTO;
 import org.example.simplemoneytransfer.dto.BankModifyDTO;
 import org.example.simplemoneytransfer.entity.Bank;
+import org.example.simplemoneytransfer.exception.BankNotFoundException;
 import org.example.simplemoneytransfer.service.BankService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -26,36 +29,46 @@ public class BankController {
     private ModelMapper modelMapper;
     
     @GetMapping(path="/banks")
-    public List<BankDTO> getAllBanks()
+    public ResponseEntity<List<BankDTO>> getAllBanks()
     {
-        return bankService.getAllBanks().stream()
+        return  new ResponseEntity<>(bankService.getAllBanks().stream()
         		                         .map(b -> modelMapper.map(b, BankDTO.class))
-        		                         .collect(Collectors.toList());
+        		                         .collect(Collectors.toList()), HttpStatus.OK);
     }
     
     @GetMapping(path="banks/{id}")
-    public BankDTO getBankById(@PathVariable Long id)
+    public ResponseEntity<BankDTO> getBankById(@PathVariable Long id)
     {
-    	 return modelMapper.map(bankService.getBankById(id), BankDTO.class);
+    	if (bankService.getBankById(id) == null)
+       		throw new BankNotFoundException(id);
+    	
+    	 return new ResponseEntity<>(modelMapper.map(bankService.getBankById(id), BankDTO.class), HttpStatus.OK);
     }
 
     @PostMapping(path="/banks")
-    public BankDTO addBank(@RequestBody BankModifyDTO newBank)
+    public ResponseEntity<BankDTO> addBank(@RequestBody BankModifyDTO newBank)
     {
-        return modelMapper.map(bankService.addBank(modelMapper.map(newBank, Bank.class)), 
-        		               BankDTO.class);
+        return new ResponseEntity<>(modelMapper.map(bankService.addBank(modelMapper.map(newBank, Bank.class)), 
+        		               BankDTO.class), HttpStatus.CREATED);
     }
 
     @PutMapping(path="/banks/{id}")
-    public BankDTO updateBank(@RequestBody BankModifyDTO updatedBank, @PathVariable Long id)
+    public ResponseEntity<BankDTO> updateBank(@RequestBody BankModifyDTO updatedBank, @PathVariable Long id)
     {
-        return modelMapper.map(bankService.updateBank(modelMapper.map(updatedBank, Bank.class), id), 
-	               BankDTO.class);
+    	if (bankService.getBankById(id) == null)
+       		throw new BankNotFoundException(id);
+    	
+        return new ResponseEntity<>(modelMapper.map(bankService.updateBank(modelMapper.map(updatedBank, Bank.class), id), 
+	               BankDTO.class), HttpStatus.OK);
     }
 
     @DeleteMapping(path="/banks/{id}")
-    public void deleteBank(@PathVariable Long id)
+    public ResponseEntity<Void> deleteBank(@PathVariable Long id)
     {
-        bankService.deleteBank(id);
+    	if (bankService.getBankById(id) == null)
+       		throw new BankNotFoundException(id);
+    	
+    	bankService.deleteBank(id);
+    	return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
